+++ R/get_posterior.R
@@ -310,10 +310,18 @@ add_attributes1 <- function(obj, dimension = c("i", "ij", "ijk"),
 
 add_attributes2 <- function(obj, covariate, dimnames_y, type) {
 
+  has_intercept_only <- function(covariate) {
+    if (is.array(covariate)) {
+      return(dim(covariate)[length(dim(covariate))] == 1)
+    } else {
+      return(identical(covariate, 1))
+    }
+  }
+
   if (type == "samples") {
     attr(obj, "dimension") <- c("Sample", "Species", "Effects")
 
-    if (identical(covariate, 1)) {
+    if (has_intercept_only(covariate)) {
       effect_name <- "(Intercept)"
     } else {
       effect_name <- dimnames(covariate)[[length(dim(covariate))]]
@@ -337,7 +345,7 @@ add_attributes2 <- function(obj, covariate, dimnames_y, type) {
   if (type == "summary") {
     attr(obj, "dimension") <- c("Species", "Effects")
 
-    if (identical(covariate, 1)) {
+    if (has_intercept_only(covariate)) {
       effect_name <- "(Intercept)"
     } else {
       effect_name <- dimnames(covariate)[[length(dim(covariate))]]